Patch.go refactoring
- Created the pkg/inject package to hold the new injection shared lib.
- Moved patch.go there.
- Added the ability to add pod labels and annotations without having to
specify the already existing ones. For that purpose, added `patch.addPodLabelsRoot()` and `addPodAnnotationsRoot()`, and repurposed `addPodLabels()` into `addPodLabel()`, and `addPodAnnotations()` into  `addPodAnnotation()`
- Removed `addDeploymentLabels()` because I couldn't find the need for
it. At least `linkerd inject` wasn't setting those labels. Please
confirm in case I'm wrong.

Still to-do: have `patch_test.go` depend on fixtures separate from the
ones under `controller/proxy-injector`

Ref #1748
